Remove condition to required review count
This one temporarily removes the condition for review count.

The check against old value is somewhat broken as it checks against
old version of the yaml file, and if the previous value was wrong
(for example because of failure in other steps) - the cached yaml
might not store the actual value.

The check should be done against the current value retrieved from
github.