Add note about GitHub Actions `app_id`
Most repositories use at least the "GitHub Actions" GitHub App.

The `app_id` of "GitHub Actions" is `15368` and that value seems to work in `logging-log4j2`. Since from a user perspective, it's hard to find out the identifier of a GitHub App, it would be probably useful to list all the ids of applications installed at the `apache` organization level.