Add warning about required checks
Typos in the names and application ids of required checks for the default branch can brick the GitHub interface.
Sure, the error can be corrected through GitBox, but for transparency reasons, it would be better never to resort to that.

This change adds a warning to test the required checks names on a **non** default branch first. If an error is introduced, it is always possible to modify the `.asf.yaml` file through GitHub.